Final commit
Local version of code & github version of code became seperated, updaing local code with github version and then reuploading small local changes to github version
