+++ getUsersAndPosts.js
@@ -5,29 +5,6 @@ function twitterAccount() {
 
 }
 
-function twitterPost(id, done) {
-    fetch(`https://api.twitter.com/1.1/statuses/show.json?id=${id}&tweet_mode=extended`, {
-        method: "get",
-        headers:  {
-            'Content-Type': 'application/json',
-            "authorization": `Bearer ${process.env.TWITTER_BEARER_TOKEN}`,
-        }
-    }).then(res => res.json()).then(post => {
-        const results = {
-            success: true,
-            post,
-        }
-        done(results);
-    }).catch((err) => {
-        console.log(err);
-        const results = {
-            success: false,
-            post,
-        }
-        done(results);
-    })
-}
-
 function linkedinAccount() {
 
 }
@@ -47,7 +24,6 @@ function facebookPost() {
 
 module.exports = {
     twitterAccount,
-    twitterPost,
     linkedinAccount,
     linkedinPost,
     facebookAccount,
