+++ src/app/my-feed/my-feed.component.html
@@ -32,7 +32,6 @@
 <!--
 <app-linkedin-account></app-linkedin-account>
 <app-linkedin-post></app-linkedin-post>
-
 <app-facebook-account></app-facebook-account>
 <app-facebook-post></app-facebook-post>
 -->
