+++ src/app/twitter-account/twitter-account.component.html
@@ -1,1 +1,3 @@
-<p>twitter-account works!</p>
+<div class="preview-background">
+    <div class="account-preview">{{ account }}</div>
+</div>
